Move Atom into java lib for deletion
