+++ radixdlt-java/radixdlt-java/src/main/java/com/radixdlt/client/core/atoms/Addresses.java
@@ -18,8 +18,6 @@
 
 package com.radixdlt.client.core.atoms;
 
-import com.radixdlt.atom.Atom;
-
 import com.radixdlt.atommodel.tokens.TokensParticle;
 import com.radixdlt.atommodel.validators.ValidatorParticle;
 import com.radixdlt.constraintmachine.Particle;
