+++ radixdlt-java/radixdlt-java/src/main/java/com/radixdlt/client/core/ledger/AtomObservation.java
@@ -22,7 +22,7 @@
 
 package com.radixdlt.client.core.ledger;
 
-import com.radixdlt.atom.Atom;
+import com.radixdlt.client.core.atoms.Atom;
 import com.radixdlt.client.core.ledger.AtomEvent.AtomEventType;
 
 public final class AtomObservation {
