+++ radixdlt-java/radixdlt-java/src/main/java/com/radixdlt/client/core/ledger/InMemoryAtomStore.java
@@ -24,7 +24,7 @@ package com.radixdlt.client.core.ledger;
 
 import com.radixdlt.atom.Substate;
 import com.radixdlt.atom.TxLowLevelBuilder;
-import com.radixdlt.atom.Atom;
+import com.radixdlt.client.core.atoms.Atom;
 import com.radixdlt.client.core.atoms.Addresses;
 import com.radixdlt.constraintmachine.Spin;
 import io.reactivex.Observable;
