+++ radixdlt-java/radixdlt-java/src/main/java/com/radixdlt/client/core/ledger/InMemoryAtomStoreReducer.java
@@ -24,7 +24,7 @@ package com.radixdlt.client.core.ledger;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
-import com.radixdlt.atom.Atom;
+import com.radixdlt.client.core.atoms.Atom;
 import com.radixdlt.client.core.atoms.Addresses;
 import com.radixdlt.client.core.atoms.AtomStatus;
 import com.radixdlt.client.core.network.RadixNodeAction;
